Chaning size_t to uint64_t

+++ src/libethash/internal.c
@@ -36,13 +36,13 @@
 #include "sha3.h"
 #endif // WITH_CRYPTOPP
 
-size_t ethash_get_datasize(const uint32_t block_number) {
+uint64_t ethash_get_datasize(const uint32_t block_number) {
     if (block_number / EPOCH_LENGTH >= 2048)
         return 0;
     return dag_sizes[block_number / EPOCH_LENGTH];
 }
 
-size_t ethash_get_cachesize(const uint32_t block_number) {
+uint64_t ethash_get_cachesize(const uint32_t block_number) {
     if (block_number / EPOCH_LENGTH >= 2048)
         return 0;
     return cache_sizes[block_number / EPOCH_LENGTH];
